Moved cli formater to proper folder
