+++ src/formats/json.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = (info) => {
+  return JSON.stringify(info);
+};
